Organize classes into package
