+++ src/main/java/airy/Deadline.java
@@ -1,48 +1,46 @@
+package airy;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-
 /**
  * This class is a subclass from task
- * A task with a starting date and an end date
+ * A task with a deadline
  */
-public class Event extends Task {
-    private final LocalDate startDate;
-    private final LocalDate endDate;
+public class Deadline extends Task {
+    private final LocalDate dueDate;
 
     private static final DateTimeFormatter INPUT_FORMAT = DateTimeFormatter.ofPattern("yyyy-MM-dd");
     private static final DateTimeFormatter OUTPUT_FORMAT = DateTimeFormatter.ofPattern("MMM dd yyyy");
 
     /**
-     * This is my event constructor
+     * This is my deadline constructor
      */
-    public Event(String taskName, String startDate, String endDate) {
+    public Deadline(String taskName, String dueDate) {
         super(taskName);
         try {
-            this.startDate = LocalDate.parse(startDate, INPUT_FORMAT);
-            this.endDate = LocalDate.parse(endDate, INPUT_FORMAT);
+            this.dueDate = LocalDate.parse(dueDate, INPUT_FORMAT);
         } catch (DateTimeParseException e) {
             throw new AiryException("Invalid date format, please use yyyy-MM-dd, e.g. 2025-06-02");
         }
     }
 
     /**
-     * Provides info of the event task in string format
+     * Provides info of the deadline task in string format
      */
     @Override
     public String toString() {
-        return String.format("[E]%s (from: %s to: %s)",
+        return String.format("[D]%s (by: %s)",
                 super.toString(),
-                startDate.format(OUTPUT_FORMAT),
-                endDate.format(OUTPUT_FORMAT));
+                dueDate.format(OUTPUT_FORMAT));
     }
 
     /**
-     * Returns the startDate and endDate in input format for Storage
+     * Returns the dueDate in input format for Storage
      */
     @Override
     public String getExtraDetailsForStorage() {
-        return startDate.format(INPUT_FORMAT) + " | " + endDate.format(INPUT_FORMAT);
+        return dueDate.format(INPUT_FORMAT);
     }
 }
