+++ src/main/java/airy/Task.java
@@ -1,3 +1,5 @@
+package airy;
+
 /**
  * Task object to keep track of task name and status
  */
