+++ src/main/java/airy/Ui.java
@@ -1,3 +1,5 @@
+package airy;
+
 import java.util.ArrayList;
 
 /**
